requirements.txt
pyyaml==6.0.1
pytest==8.3.2
groq==0.9.0
